некоторые изменения в репозиториях, начал настройку локализации

+++ src/main/java/com/samsolutions/service/RoleService.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Service;
 
 @Service
 public interface RoleService {
-    Role findRoleById(int id);
+    Role findRoleById(long id);
 }
